Created Front Page
Took a long time to plan out what I wanted. I noticed that I took a lot
of time when I didn't have the resources I needed or if I didn't know
what I wanted to look like. I think that HTML/CSS will be 70% planning
30% coding. I'm definitely not a designer yet I hope to make this look
as best as it can.

To Do:
1) Add logo.png
2) Create other pages
3) Space menu items properly
3) Space box items properly