Attempted footer and account.html
I attempted to make the footer a bit more easy to use. I added the
navbar to the other pages, including the new <head> and such. I fixed up
the account.html page and now it can be used as a form to get started. I
wish to edit the footer and get the other pages to a point where I can
go to JavaScript stuff.

Here are the steps going forward:

Bootstrap -> Mobile first
Ability to make an account and log in
List of clubs page (books  and start dates)
  This part will include databasing the books and having users be  able
to add there own clubs
Calendar and messanger functionalities
  This is the MVP (minimmum valuable project)
Include content for learn page (what is QBC?)