Template page, JavaScript, Node.js
I've gotten to my first JavaScript code! Time for a party! I tighten up
the form on the account.html page and then allow for data to have going
inputed (if Node is running). The data isn't going into MongoDB, yet
that's the next step. I'm not too focused on design right now, I want to
get the functionality in. I also learned that I can have two
XMLHttpRequest JavaScript objects on the same page as long as they are
in different objects...hm...